Argon2Config: disallow null salt
Section 3.1 of the spec https://www.password-hashing.net/argon2-specs.pdf specfies a minimum length of 8 for the salt. The reference implementation does not allow empty salts either. Although line 423 in core.c looks like it allows a null salt if the length is zero, the zero length will then actually be rejected by the condition in line 428.

This change makes Argon2Config generate a random salt with the recommended default length of 16 when the getter of the `Salt` property is called without setting a salt beforehand. This allows the property to be non-nullable, and cases dealing with a spec-violating null salt be removed.

Removal of the null-handling case also gets rid of an erroneous early bailout in that case from the encoder. That had the effect of omitting the hash from the encoded string when no salt was given (#52). Note that just skipping the salt would not work, because the decoder would then interpret the hash as salt instead.

The test in ThreadsDontMatter.cs had to be changed so that both configs use the same salt. Note that before the fix that test didn't actually test anything because of the bug mentioned above: an empty salt would cause the hash to be omitted from the encoded string as well.

This fixes https://github.com/mheyman/Isopoh.Cryptography.Argon2/issues/52